Graph zoom: Use float variables when calculating the upper/lower limits of the graph
git-svn-id: svn://svn.code.sf.net/p/xymon/code@6608 44351d6e-118b-4698-b696-ce33095ecaa4

+++ branches/4.3.0/xymond/webfiles/zoom.js
@@ -587,8 +587,8 @@ function onMouseUpEvent(e) {
 
  graphStart = parseInt(gUrlObj.getUrlParameterValue("graph_start"));
  graphEnd = parseInt(gUrlObj.getUrlParameterValue("graph_end"));
- graphTop = parseInt(gUrlObj.getUrlParameterValue("upper"));
- graphBottom = parseInt(gUrlObj.getUrlParameterValue("lower"));
+ graphTop = parseFloat(gUrlObj.getUrlParameterValue("upper"));
+ graphBottom = parseFloat(gUrlObj.getUrlParameterValue("lower"));
  haveGraphLimits = (gUrlObj.getUrlParameterValue("upper") != undefined) && (gUrlObj.getUrlParameterValue("lower") != undefined);
 
  idxStr = "";
